django_cards/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
backcall==0.1.0
bleach==3.0.2
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
cloudpickle==0.6.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dask==1.0.0
dataclasses==0.6
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
djongo==1.2.31
entrypoints==0.2.3
enum34==1.1.6
httplib2==0.9.2
idna==2.6
imageio==2.4.1
imgaug==0.2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja==22.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.2
notebook==5.7.4
numpy==1.15.4
oauth==1.0.1
olefile==0.45.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
pip==18.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.0.0
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
pygobject==3.26.1
pymacaroons==0.13.0
pymongo==3.7.2
PyNaCl==1.1.2
pyparsing==2.3.0
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.0.1
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.14.1
scipy==1.1.0
SecretStorage==2.3.1
Send2Trash==1.5.0
setuptools==39.0.1
Shapely==1.6.4.post2
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
system-service==0.3
systemd-python==234
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib==31.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xkit==0.0.0
zope.interface==4.3.2
